+++ Source/Node.cpp
@@ -28,9 +28,9 @@ void Node::draw(SDL_Renderer* renderer) {
 void Node::draw(SDL_Renderer* renderer, SDL_Color color){
     SDL_RenderDrawLine(renderer, m_Inner, color);
     SDL_RenderDrawLine(renderer, m_Outer, color);
-    SDL_RenderDrawLine(renderer, m_Inner->m_pBegin->x, m_Inner->m_pBegin->y, m_Outer->m_pBegin->x, m_Outer->m_pBegin->y, color);
-    SDL_RenderDrawLine(renderer, m_Inner->m_pEnd->x, m_Inner->m_pEnd->y, m_Outer->m_pEnd->x, m_Outer->m_pEnd->y, color);
-//    SDL_RenderDrawLine(renderer, m_Axis, color);
+    SDL_RenderDrawLine(renderer, (int)m_Inner->m_pBegin->x, (int)m_Inner->m_pBegin->y, (int)m_Outer->m_pBegin->x, (int)m_Outer->m_pBegin->y, color);
+    SDL_RenderDrawLine(renderer, (int)m_Inner->m_pEnd->x, (int)m_Inner->m_pEnd->y, (int)m_Outer->m_pEnd->x, (int)m_Outer->m_pEnd->y, color);
+
 }
